LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-anchor-positioning-005-expected.txt


FAIL CSS Content Visibility: auto not skipped, but skipped after change of positioned-element style assert_equals: #positioned should be anchored because #anchor is on-screen. expected 209 but got 8
